+++ CHANGELOG.md
@@ -31,6 +31,7 @@ All notable changes to this project will be documented in this file. This change
   visually distinct.
 - Parts of the introductory walk-through in `README.md` had become
   stale and needed to be updated.
+- The API documentation for `patch-fixture!` was fleshed out.
 
 ## [0.1.4] - 2015-09-27
 ### Added